tidy fix

+++ tasks/lukin_i_elem_vec_sum/common/include/common.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 #include <string>
 #include <tuple>
 #include <vector>
@@ -10,7 +11,7 @@
 namespace lukin_i_elem_vec_sum {
 
 using InType = std::vector<int>;
-using OutType = long long;
+using OutType = int64_t;
 using TestType = std::tuple<size_t, std::string>;
 using BaseTask = ppc::task::Task<InType, OutType>;
